Split the logic into two classes and a base class
One class is for the unbounded version and the other is for the fixed size version
