+++ amulet/api/partial_3d_array/__init__.py
@@ -1,1 +1,2 @@
-from .partial_3d_array import Partial3DArray
+from .unbounded_partial_3d_array import UnboundedPartial3DArray
+from .bounded_partial_3d_array import BoundedPartial3DArray
