+++ amulet/api/partial_3d_array/base_partial_3d_array.py
@@ -1,6 +1,5 @@
-from typing import Optional, Dict, Union, Tuple, Generator, overload, Iterable, Type
+from typing import Optional, Dict, Union, Tuple, Type
 import numpy
-import math
 
 from .util import (
     get_sliced_array_size,
@@ -9,26 +8,13 @@ from .util import (
 )
 
 
-class Partial3DArray:
-    """This is designed to work similarly to a numpy.ndarray but stores the data in a very different way.
-    A numpy.ndarray stores a fixed size continuous array which for large arrays can become unmanageable.
-    Sparse arrays allow individual values to exist which can be great where a small set of values are
-    defined in a large area but can be less efficient in some cases than a continuous array.
-
-    This class was born out of the need for an array that has a fixed size in the horizontal distances but an
-    unlimited height in the vertical distance (both above and below the origin)
-    This is achieved by splitting the array into sections of fixed height and storing them sparsely so that only
-    the defined sections need to be held in memory.
-
-    This class also implements an API that resembles that of the numpy.ndarray but it is not directly compatible with numpy.
-    Chained indexing does not currently work as __getitem__ returns a copy of the array data rather than a view into it.
-    To use numpy fully the section data will need to be directly used.
-    """
+class BasePartial3DArray:
+    """Do not use this class directly. Use UnboundedPartial3DArray or BoundedPartial3DArray"""
 
     @classmethod
     def from_partial_array(
             cls,
-            parent_array: "Partial3DArray",
+            parent_array: "BasePartial3DArray",
             start: Tuple[int, Optional[int], int],
             stop: Tuple[int, Optional[int], int],
             step: Tuple[int, int, int]
@@ -72,7 +58,7 @@ class Partial3DArray:
             start: Tuple[Optional[int], Optional[int], Optional[int]],
             stop: Tuple[Optional[int], Optional[int], Optional[int]],
             step: Tuple[Optional[int], Optional[int], Optional[int]],
-            parent_array: Optional["Partial3DArray"] = None,
+            parent_array: Optional["BasePartial3DArray"] = None,
             sections: Optional[Dict[int, numpy.ndarray]] = None
     ):
         self._dtype = dtype
@@ -100,7 +86,7 @@ class Partial3DArray:
                 assert section.shape == self._section_shape, "The size of all sections must be equal to the section_shape."
                 assert section.dtype == self._dtype, "The given dtype does not match the arrays given."
 
-        elif isinstance(parent_array, Partial3DArray):
+        elif isinstance(parent_array, BasePartial3DArray):
             # populate from the array
             assert parent_array.section_shape == self._section_shape, "The parent section shape must match the given section_shape."
             assert parent_array.dtype == self._dtype, "The parent dtype must match the given dtype"
@@ -176,45 +162,19 @@ class Partial3DArray:
     def section_shape(self) -> Tuple[int, int, int]:
         return self._section_shape
 
-    # def __array__(self):
-    #     raise NotImplementedError  # todo
+    def __array__(self):
+        raise NotImplementedError
 
     def __contains__(self, item: int):
         return item in self._sections
 
     def __iter__(self):
-        raise NotImplementedError(
+        raise Exception(
             "Please use sections method if this is what you are trying to achieve"
         )
 
-    @property
-    def sections(self) -> Iterable[int]:
-        """An iterable of the section indexes that exist"""
-        return self._sections.keys()
+    def __getitem__(self, item):
+        raise NotImplementedError
 
-    def create_section(self, cy: int):
-        if self._parent_array is None:
-            self._sections[cy] = numpy.full(
-                self.section_shape,
-                self.default_value,
-                dtype=self._dtype
-            )
-        else:
-            self._parent_array.create_section(cy)
-
-    def add_section(self, cy: int, section: numpy.ndarray):
-        if self._parent_array is None:
-            assert section.shape == self._section_shape, "The size of all sections must be equal to the section_shape."
-            assert section.dtype == self._dtype, "the given dtype does not match the arrays given."
-        else:
-            self._parent_array.add_section(cy, section)
-
-    def get_section(self, cy: int) -> numpy.ndarray:
-        """Get the section ndarray for a given section index.
-        :param cy: The section y index
-        :return: Numpy array for this section
-        :raises: KeyError if no section exists with this index
-        """
-        if cy not in self._sections:
-            self.create_section(cy)
-        return self._sections[cy]
+    def __setitem__(self, key, value):
+        raise NotImplementedError
