+++ amulet/api/partial_3d_array/bounded_partial_3d_array.py
@@ -1,0 +1,13 @@
+from .base_partial_3d_array import BasePartial3DArray
+
+
+class BoundedPartial3DArray(BasePartial3DArray):
+    """This class should behave the same as a numpy array in all three axis
+    but the data internally is stored in chunks to minimise memory usage.
+    The array has a fixed size in all three axis much like a numpy array."""
+
+    def __getitem__(self, item):
+        raise NotImplementedError
+
+    def __setitem__(self, key, value):
+        raise NotImplementedError
